1.20.1 Update
Perhaps the most significant change was the migration of the previous 4 slot system to the now 9 slot + slot binding system like the one used in project korra! Each new slot corresponds to a position in your hotbar (1-9). To use a move, select the corresponding slot and then punch, right-click, or double sneak. You can customize each slot with different activation methods. Hovering over a slot shows the associated move, and performing the correct action triggers the move. Here's an example: Let's say you learn an move called "Air Blast" and want to use it. Open the scroll menu, go to "Slot Choice," and select Slot 1. Then choose "Air Blast" from the options and set the activation method to "Sneak & Punch." Submit your selection, and you'll see a chat message confirming it. To use the move, select the first hotbar slot and sneak and punch at the same time. The same can be done for every other slot, and you can activate them all differently. Slot 1 could be right click, because you always have a sword there, and slot 5 could be double sneak. That's a huge change, but it's not the only one! There also, new animations for almost every single move (with the exception of 4 special moves), 64 unlockable new moves and passives for every single bending style using the skill tree, more sub-bending moves for the benders that only had one or two, and a ton of unique new admin features for moderating servers and realms.